Update partners.yml
Ack -- I overwrote some of your changes, @seanherron. I added back in the instructions at the top, but one thing I noticed: 

You had:
# status: implemented OR company

But we'd talked about only having "committed" as a status (otherwise, it was implemented).

So I redid it as:

# This file outlines all of the partners listed on the site. Below is an example item:
  
# - name: "Name of Partner"
#   url: "http://partner.com"
#   type: utility OR company
#   status: IF committed
#   featured: true (if item should be featured)

Hope this works -- so sorry if I broke something.